requirements.txt
PyYAML==6.0.1
pyparsing==3.0.8
clickhouse_connect==0.7.8
mysql-connector-python==8.3.0
